python-parser/requirements.txt
flask
pdfplumber==0.11.0
PyMuPDF==1.24.1
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
pandas==2.0.3
numpy==1.24.4
python-dateutil==2.9.0
regex==2024.4.28
requests==2.32.2
pdfminer.six
camelot-py[cv]
PyPDF2
gunicorn
